Refactor cli workspace
